moved InteractiveMarkerGraphHandle to a separate file.
